ml-platform/examples/ray-dataprocessing/src/requirements.txt
ray==2.7.1
ray[client]==2.7.1
spacy==3.7.4
google-cloud-storage==2.16.0
pandas==2.2.1
gcsfs==2024.3.1
fsspec==2024.3.1
